create a separate aug base
